feat2 : Scene Updating system added, collisions are now managed by the scene for each registered GameObject + tests + Makefile update

+++ include/GameMap.h
@@ -7,13 +7,13 @@
 class GameMap {
     public:
         void Render() {
-            for(int i = 0; i < m_MapLayers.size(); i++) {
+            for(size_t i = 0; i < m_MapLayers.size(); i++) {
                 m_MapLayers[i]->Render();
             }
         }
 
         void Update() {
-            for (int i = 0; i < m_MapLayers.size(); i++) {
+            for (size_t i = 0; i < m_MapLayers.size(); i++) {
                 m_MapLayers[i]->Update();
             }
         }
